expression: move "Constant" to a standalone source file
