+++ expression/expression.go
@@ -14,7 +14,6 @@
 package expression
 
 import (
-	"bytes"
 	goJSON "encoding/json"
 	"fmt"
 
@@ -25,7 +24,6 @@ import (
 	"github.com/pingcap/tidb/mysql"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/terror"
-	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/types"
 	"github.com/pingcap/tidb/util/types/json"
 )
@@ -272,156 +270,6 @@ func evalExprToJSON(expr Expression, row []types.Datum, _ *variable.StatementCon
 		return val.GetMysqlJSON(), false, nil
 	}
 	panic(fmt.Sprintf("cannot get JSON result from %s expression", types.TypeStr(expr.GetType().Tp)))
-}
-
-// One stands for a number 1.
-var One = &Constant{
-	Value:   types.NewDatum(1),
-	RetType: types.NewFieldType(mysql.TypeTiny),
-}
-
-// Zero stands for a number 0.
-var Zero = &Constant{
-	Value:   types.NewDatum(0),
-	RetType: types.NewFieldType(mysql.TypeTiny),
-}
-
-// Null stands for null constant.
-var Null = &Constant{
-	Value:   types.NewDatum(nil),
-	RetType: types.NewFieldType(mysql.TypeTiny),
-}
-
-// Constant stands for a constant value.
-type Constant struct {
-	Value   types.Datum
-	RetType *types.FieldType
-}
-
-// String implements fmt.Stringer interface.
-func (c *Constant) String() string {
-	return fmt.Sprintf("%v", c.Value.GetValue())
-}
-
-// MarshalJSON implements json.Marshaler interface.
-func (c *Constant) MarshalJSON() ([]byte, error) {
-	buffer := bytes.NewBufferString(fmt.Sprintf("\"%s\"", c))
-	return buffer.Bytes(), nil
-}
-
-// Clone implements Expression interface.
-func (c *Constant) Clone() Expression {
-	con := *c
-	return &con
-}
-
-// GetType implements Expression interface.
-func (c *Constant) GetType() *types.FieldType {
-	return c.RetType
-}
-
-// GetTypeClass implements Expression interface.
-func (c *Constant) GetTypeClass() types.TypeClass {
-	return c.RetType.ToClass()
-}
-
-// Eval implements Expression interface.
-func (c *Constant) Eval(_ []types.Datum) (types.Datum, error) {
-	return c.Value, nil
-}
-
-// EvalInt returns int representation of Constant.
-func (c *Constant) EvalInt(_ []types.Datum, sc *variable.StatementContext) (int64, bool, error) {
-	if c.GetType().Tp == mysql.TypeNull {
-		return 0, true, nil
-	}
-	val, isNull, err := evalExprToInt(c, nil, sc)
-	return val, isNull, errors.Trace(err)
-}
-
-// EvalReal returns real representation of Constant.
-func (c *Constant) EvalReal(_ []types.Datum, sc *variable.StatementContext) (float64, bool, error) {
-	if c.GetType().Tp == mysql.TypeNull {
-		return 0, true, nil
-	}
-	val, isNull, err := evalExprToReal(c, nil, sc)
-	return val, isNull, errors.Trace(err)
-}
-
-// EvalString returns string representation of Constant.
-func (c *Constant) EvalString(_ []types.Datum, sc *variable.StatementContext) (string, bool, error) {
-	if c.GetType().Tp == mysql.TypeNull {
-		return "", true, nil
-	}
-	val, isNull, err := evalExprToString(c, nil, sc)
-	return val, isNull, errors.Trace(err)
-}
-
-// EvalDecimal returns decimal representation of Constant.
-func (c *Constant) EvalDecimal(_ []types.Datum, sc *variable.StatementContext) (*types.MyDecimal, bool, error) {
-	if c.GetType().Tp == mysql.TypeNull {
-		return nil, true, nil
-	}
-	val, isNull, err := evalExprToDecimal(c, nil, sc)
-	return val, isNull, errors.Trace(err)
-}
-
-// EvalTime returns DATE/DATETIME/TIMESTAMP representation of Constant.
-func (c *Constant) EvalTime(_ []types.Datum, sc *variable.StatementContext) (val types.Time, isNull bool, err error) {
-	if c.GetType().Tp == mysql.TypeNull {
-		return val, true, nil
-	}
-	val, isNull, err = evalExprToTime(c, nil, sc)
-	return val, isNull, errors.Trace(err)
-}
-
-// EvalDuration returns Duration representation of Constant.
-func (c *Constant) EvalDuration(_ []types.Datum, sc *variable.StatementContext) (val types.Duration, isNull bool, err error) {
-	if c.GetType().Tp == mysql.TypeNull {
-		return val, true, nil
-	}
-	val, isNull, err = evalExprToDuration(c, nil, sc)
-	return val, isNull, errors.Trace(err)
-}
-
-// EvalJSON returns JSON representation of Constant.
-func (c *Constant) EvalJSON(_ []types.Datum, sc *variable.StatementContext) (json.JSON, bool, error) {
-	val, isNull, err := evalExprToJSON(c, nil, sc)
-	return val, isNull, errors.Trace(err)
-}
-
-// Equal implements Expression interface.
-func (c *Constant) Equal(b Expression, ctx context.Context) bool {
-	y, ok := b.(*Constant)
-	if !ok {
-		return false
-	}
-	con, err := c.Value.CompareDatum(ctx.GetSessionVars().StmtCtx, y.Value)
-	if err != nil || con != 0 {
-		return false
-	}
-	return true
-}
-
-// IsCorrelated implements Expression interface.
-func (c *Constant) IsCorrelated() bool {
-	return false
-}
-
-// Decorrelate implements Expression interface.
-func (c *Constant) Decorrelate(_ *Schema) Expression {
-	return c
-}
-
-// HashCode implements Expression interface.
-func (c *Constant) HashCode() []byte {
-	var bytes []byte
-	bytes, _ = codec.EncodeValue(bytes, c.Value)
-	return bytes
-}
-
-// ResolveIndices implements Expression interface.
-func (c *Constant) ResolveIndices(_ *Schema) {
 }
 
 // composeConditionWithBinaryOp composes condition with binary operator into a balance deep tree, which benefits a lot for pb decoder/encoder.
